% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_trelliscope.R
\name{cluster_trelliscope}
\alias{cluster_trelliscope}
\title{cluster_trelliscope}
\usage{
cluster_trelliscope(data, trans = "identity", name = "Cluster Results",
  group = "common", path = "~/trelliscope", selfContained = FALSE)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count
* Prediction}

\item{trans}{Scale of the y-axis, see `ggplot2::scale_y_continuous()`
* Default is `identity`}

\item{name}{Name of the trelliscope view
* Default is `Cluster Results`}

\item{group}{Group of trelliscope views
* Default is `common`}

\item{path}{Base directory of trelliscope app
* Default is `~/trelliscope`}

\item{selfContained}{Create the individual pre-rendered panels
* Default is `FALSE`}
}
\value{
A trelliscopejs object
}
\description{
**Intented to take data in long format for trelliscopejs use**
}
\examples{
test
}
\seealso{
* `ggplot2::scale_y_continuous()`
* [`trelliscopejs`](https://hafen.github.io/trelliscopejs/)
}
