% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point.R
\name{cut_point}
\alias{cut_point}
\title{cut_point}
\usage{
cut_point(data, movAvg = 21, startQuant = 0.2, movAvg_zeroDays = 1,
  dialBack = 0)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{movAvg}{Moving average 
* Default is `21`}

\item{startQuant}{Quantile percentage to determine start date cut_point
* Default is `0.2`}

\item{movAvg_zeroDays}{Number of consecetive moving averages evaluated at zero to begin end date cut_point process
* Defaults to `1`}

\item{dialBack}{Number of days to 'dialBack' to determine a new end date cutPoint
* Defaults to `0`}
}
\value{
filtered tibble with `cut_point` algorithim performed
}
\description{
**Designed to implement `cut_point` algorithin on column**
}
\examples{
test
}
