% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point_trelliscope.R
\name{cut_point_trelliscope}
\alias{cut_point_trelliscope}
\title{cut_point_trelliscope}
\usage{
cut_point_trelliscope(data, movAvg = 21, trans = "identity",
  name = "cut_point Results", group = "common",
  path = "~/trelliscope", selfContained = F)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{movAvg}{Moving average 
* Default is `21`}

\item{trans}{Scale of the y-axis, see `ggplot2::scale_y_continuous()`
* Default is `identity`}

\item{name}{Name of the trelliscope view
* Default is `cut_point Results`}

\item{group}{Group of trelliscope views
* Default is `common`}

\item{path}{Base directory of trelliscope app
* Default is `~/trelliscope`}

\item{selfContained}{Create the individual pre-rendered panels
* Default is `FALSE`}
}
\value{
trelliscopejs object
}
\description{
**Intended to take long data and output trelliscope with `cut_point` vertical lines**
}
\examples{
test
}
