% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{do_interval_work}
\alias{do_interval_work}
\title{do_interval_work}
\usage{
do_interval_work(data, type, intervalLength, divide)
}
\arguments{
\item{data}{Non-nested tibble with the following columns:
* A unique identifier (usually made by a `group_by`` call)
* A continious date column
* A numeric column, one observation per date per unique identifier}

\item{type}{Must be one of following strings:
* `years`
* `months`
* `weeks`
* `days`}

\item{intervalLength}{Must be positive integer excluding zero}

\item{divide}{Must be one of following strings:
* `A`
* `R`
* `L`}
}
\value{
List that contains [nestedTibble, tibble]
}
\description{
**Designed to create cognostics by certain interval/ratio in the past**
}
\examples{
NULL
}
