% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{do_nested_interval_columns_work}
\alias{do_nested_interval_columns_work}
\title{do_nested_interval_columns_work}
\usage{
do_nested_interval_columns_work(data, type, interval, divide)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{type}{Must be one of following strings:
* `years`
* `months`
* `weeks`
* `days`}

\item{interval}{Must be positive integer excluding zero}

\item{divide}{Must be one of following strings:
* `A`
* `R`
* `L`}
}
\value{
nested tibble
}
\description{
**Designed to create cognostics by certain interval/ratio in the past**
}
\examples{
test
}
