% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{extract_nested_columns}
\alias{extract_nested_columns}
\title{extract_nested_columns}
\usage{
extract_nested_columns(data)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}
}
\value{
unnested tibble
}
\description{
**Designed to unnest a nested tibble, which is a list of lists**
}
\examples{
test
}
