% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_LadderSequence}
\alias{find_LadderSequence}
\title{find_LadderSequence}
\usage{
find_LadderSequence(data, sequence)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{sequence}{Must be one of following strings:
* `I` - Increasing
* `D` - Decreasing
* `IP` - Increasing Positive
* `P` - Decreasing Positive
* `IN` - Increasing Negative
* `DN` - Decreasing Negative}
}
\value{
integer
}
\description{
**Inteded to find largest sequence of many ladder sequences**
}
\examples{
test
}
