% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{make_all_nested_columns}
\alias{make_all_nested_columns}
\title{make_all_nested_columns}
\usage{
make_all_nested_columns(data, append = TRUE)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}
}
\value{
nested and joined tibble
}
\description{
**Designed to create then join four nested tibbles into one tibble
with grouped variables in `data` being the identifier(s) by:**
* Daily
* Weekly
* Monthly
* Yearly
}
\examples{
test
}
