% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{make_nested_columns}
\alias{make_nested_columns}
\title{make_nested_columns}
\usage{
make_nested_columns(data, type, append = TRUE)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{type}{All meaningfull specifications in
English language are supported.
Stable arguments are:
* `day`
* `week`
* `month`
* `year`}
}
\value{
nested tibble
}
\description{
**Designed to append on numerous descriptive numeric cognostics as columns
then nest into a tibble with grouped variables in `data` being the identifier(s).
Based on `type` parameter, will aggregate and create cognostics by:**
* Day
* Week
* Month
* Year
}
\examples{
test
}
