% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{make_nested_interval_columns}
\alias{make_nested_interval_columns}
\title{make_nested_interval_columns}
\usage{
make_nested_interval_columns(complexObject, intervalType, intervalLength,
  append = TRUE)
}
\arguments{
\item{nestTib}{Nested tibble with `AccountNumber`` as identifier}

\item{rawData}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{type}{Must be one of following strings:
* `years`
* `months`
* `weeks`
* `days`}

\item{interval}{Must be positive integer excluding zero}
}
\value{
nested tibble
}
\description{
**Designed to create cognostics by certain interval/ratio in the past**
}
\examples{
test
}
