% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_core}
\alias{nest_core}
\title{nest_core}
\usage{
nest_core(data, type)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}

\item{type}{All meaningfull specifications in
English language are supported.
Stable arguments are:
* `day`
* `week`
* `month`
* `year`}
}
\value{
nested tibble
}
\description{
**Designed to append on numerous descriptive numeric cognostics as columns
then nest into a tibble with "AccountNumber" being the identifier.
Based on `type` parameter, will aggregate and create cognostics by:**
* Day
* Week
* Month
* Year
}
\examples{
test
}
