% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_todo}
\alias{nest_todo}
\title{nest_todo}
\usage{
nest_todo(data)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
* Account Number (unique identifier)
* Date
* Count}
}
\value{
nested and joined tibble
}
\description{
**Designed to create then join four nested tibbles into one tibble
with `AccountNumber` being the identifier by:**
* Daily
* Weekly
* Monthly
* Yearly
}
\examples{
test
}
