% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{overtime_by}
\alias{overtime_by}
\title{overtime_by}
\usage{
overtime_by(data, type)
}
\arguments{
\item{data}{Nested tibble or non-nested tibble with the following columns:
* A unique identifier (usually made by a `group_by`` call)
* A continious date column
* A numeric column, one observation per date per unique identifier}

\item{type}{All meaningfull specifications in English language are supported.
Stable arguments are:
* `day`
* `week`
* `month`
* `year`}
}
\value{
List that contains [nestedTibble, tibble]
}
\description{
**Designed to append on numerous summary statistics as columns
then nest into a tibble with grouped variables in `data` being the identifier(s).
Based on `type` parameter, will aggregate and create cognostics by:**
* Day
* Week
* Month
* Year
To allow further appending, the newly created nested tibble and the original data
will be return in a list.
}
\examples{
NULL
}
