% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{overtime_by_interval}
\alias{overtime_by_interval}
\title{overtime_by_interval}
\usage{
overtime_by_interval(data, intervalType, intervalLength)
}
\arguments{
\item{data}{Nested tibble or non-nested tibble with the following columns:
* A unique identifier (usually made by a `group_by`` call)
* A continious date column
* A numeric column, one observation per date per unique identifier}

\item{intervalType}{Must be one of following strings:
* `years`
* `months`
* `weeks`
* `days`}

\item{intervalLength}{Must be positive integer excluding zero}
}
\value{
List that contains [nestedTibble, tibble]
}
\description{
**By time interval, designed to append on numerous summary statistics as columns.
Then nest into a tibble with grouped variables in `data` being the identifier(s).
Based on `intervalType` and `intervalLength` parameters, will aggregate
and create columns accordingly**
To allow further appending, the newly created nested tibble and the original data
will be return in a list.
}
\examples{
NULL
}
