% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{overtime_get}
\alias{overtime_get}
\title{overtime_get}
\usage{
overtime_get(data)
}
\arguments{
\item{data}{List with the following elements:
* Nested Tibble
* Tibble}
}
\value{
Unnested tibble
}
\description{
**Return the first element in list that contains [nestedTibble, tibble]**
}
\examples{
NULL
}
