% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{overtime_unnest}
\alias{overtime_unnest}
\title{overtime_unnest}
\usage{
overtime_unnest(data)
}
\arguments{
\item{data}{Nested tibble or tibble with the following columns:
* A unique identifier (usually made by a `group_by`` call)
* A continious date column
* A numeric column, one observation per date per unique identifier}
}
\value{
Unnested tibble
}
\description{
**Designed to unnest a nested tibble, which is a list of lists**
}
\examples{
NULL
}
