% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{prepare_intervals}
\alias{prepare_intervals}
\title{prepare_intervals}
\usage{
prepare_intervals(data, intervalType, intervalLength)
}
\arguments{
\item{data}{Non-nested tibble with the following columns:
* A unique identifier (usually made by a `group_by`` call)
* A continious date column
* A numeric column, one observation per date per unique identifier}

\item{intervalType}{Must be one of following strings:
* `years`
* `months`
* `weeks`
* `days`}

\item{intervalLength}{Must be positive integer excluding zero}
}
\value{
List that contains [nestedTibble, tibble]
}
\description{
**Designed to prepare data by pulling out left, right, and ratio intervals**
}
\examples{
NULL
}
